requirements.txt
Flask>=3.0.0,<4
flask-cors>=4.0.0,<6
Flask-SQLAlchemy>=3.1.0,<4
Flask-Login>=0.6.3,<1
Flask-WTF>=1.2.0,<2
WTForms>=3.1.0,<4
email-validator>=2.0.0,<3
gunicorn>=22.0.0,<24
psycopg2-binary>=2.9.9,<3
python-dotenv>=1.0.0,<2
pytest>=8.0.0,<9
pytest-flask>=1.3.0,<2
